10mo paso, tuve que hacer mil cosas para poder crearr mi scaffolding, tenia un error el cual lo solucione c¿creando un nuevo DBContext extra en el cual sirve como apoyo para el PeiddosDBContext para que el fcore pueda crear el contexto para el scafolding, y cambie algunos valores de los models

+++ pedidos/Models/Orderitem.cs
@@ -6,7 +6,7 @@
         public int IdPedido { get; set; }
         public int IdProducto { get; set; }
         public int Cantidad { get; set; }
-        public int Subtotal { get; set; }
+        public decimal Subtotal { get; set; } 
 
         public Order Pedido { get; set; }
         public Product Producto { get; set; }
